Chunky but mechanical reworking of all the build scripts.
The main change is to eliminate nearly all of the variables in setup.sh, many of which were variables for values which are effectively constant, so they were just adding complication to the scripts.    In the process, I have looked through all the hacks again, and have a WAY better idea of how to resolve these now than I did at the time we were first doing the cross-builds.